refactor(server): rename the log package to reqctx
The package holds per-request context values (request ID, client IP), not a
logger, and its client IP feeds both the CLI log attributes and the session
audit record. reqctx names what it is; it stays a low-level package because the
auth flow reads it and cannot import server.
