+++ server/reqctx/reqctx.go
@@ -1,4 +1,4 @@
-package log
+package reqctx
 
 import (
 	"context"
